Codex continuation cleanup pass: rustfmt + clippy + node static/dynamic discriminator + doc updates
Closes the ~126 project-wide clippy issues across 5 crates flagged in
docs/CODEX_FULL_HANDOFF_2026_05_05.md as the pre-merge P1 blocker for
the next PR to main, without API refactors.

Cargo crates touched (formatting + clippy fixes only unless noted):
- agent_core/: bridge.rs, channel_relay.rs, agent_loop.rs, approval
  callers, cognitive_dag/{companions, dispatch, edge, macaroons,
  merkle, migration, resonance, storage}, cognitive_weight,
  compaction, effect/receipt, error.rs (nested-if-let safety
  refactor), etl/ffi, live_files, lsp_runtime, mcp/client,
  nightbrain/{live, mod}, oplog, providers/{claude, gemini,
  openai_compatible}, provenance/{ledger, replay}, pty,
  resources/{permissions, tool_authz}, route/{variant_b_classifiers,
  variant_c_providers}, security, skill_router,
  storage/hyperbolic_topology, tools/{chunk_reduce, cli_passthrough,
  computer_use, discovery, file_ops, web_fetch, workspace_search},
  variant_ladder
- agent_core/tests/route_salvage.rs
- agent_core/examples/generate_sample_epbundle.rs
- agent_core/src/bin/epistemos_doctrine_lint.rs
- epistemos-core/, graph-engine/, omega-ax/, omega-mcp/

Substantive additions (not pure formatting), all matching their
state:candidate briefs in CODEX_FULL_HANDOFF:
- agent_core/src/cognitive_dag/node.rs: `NodeKind::is_dynamic_rooted()`
  + `all_node_kind_variants()` test helper + new content-addressing
  determinism tests (`node_id_is_deterministic_for_identical_content`,
  `different_content_produces_different_id`). This is the
  Static/Dynamic discriminator scoping work — an O(1) classifier on
  NodeKind that says whether a node represents a content-addressed
  snapshot (Note/Claim/Evidence/Skill/Tool/Procedure/Event/Capability)
  or a dynamic-rooted artifact (Companion/Model). No call sites change
  yet — the discriminator is wired into the type but Phase 8.H mirror
  authority remains the gating sign-off before consumers light up.

JSON files updated (auto-formatted perf/signature baselines, no
threshold change):
- agent_core/perf-baselines/* (six baseline harness JSON files)
- agent_core/perf-baselines/code_token_parser_*.json
- agent_core/perf-baselines/graph_payload_construction_*.json
- agent_core/perf-baselines/markdown_parser_*.json
- agent_core/perf-baselines/graph_ffi_bridge_fixture_*.json
- agent_core/perf-baselines/mlx_thermal_policy_*.json
- agent_core/perf-baselines/true_rust_callback_*.json
- agent_core/perf-baselines/uniffi_callback_handle_*.json

Doc updates that track the cleanup status:
- docs/AGENT_PROGRESS.md (cleanup checkpoint)
- docs/APP_ISSUES_AUTO_FIX.md
- docs/A1_REDB_PERSISTENT_BACKEND_SCOPING_2026_05_05.md
- docs/B1_TAMAGOTCHI_BRAINEXPORT_LIFT_TARGETS_2026_05_05.md
- docs/B2_LIVE_FILES_AND_SUBSTRATE_LIFT_TARGETS_2026_05_05.md
- docs/B3_OBSCURA_BROWSER_LIFT_TARGETS_2026_05_05.md
- docs/CANONICAL_SWEEP_CLOSEOUT_2026_05_05.md
- docs/CD_008_PARTIAL_CLOSURE_2026_05_05.md
- docs/CODEX_FULL_HANDOFF_2026_05_05.md (continuation status section)
- docs/CODEX_VERIFICATION_HANDOFF_2026_05_05.md
- docs/RESOURCE_INVENTORY.md
- docs/STATIC_VS_DYNAMIC_WEIGHT_DELIBERATION_2026_05_05.md
- docs/fusion/EPISTEMOS_FINAL_DOCTRINE_2026_05_01.md

Verified:
- cargo build (default features): clean
- cargo build --no-default-features --features "mas-build,lsp-runtime": clean
- cargo build --no-default-features --features "pro-build,lsp-runtime": clean
- cargo test --lib: 872 passed, 0 failed
- xcodebuild -scheme Epistemos: BUILD SUCCEEDED

+++ epistemos-core/src/auto_tuner/hyperparams.rs
@@ -79,7 +79,7 @@ pub fn auto_tune(
     };
 
     let model_memory = (model_size_b * 1000.0 * 0.5) as u32;
-    let lora_memory = (lora_rank * 32 * 2 * 128 * 2 / 1024) as u32;
+    let lora_memory = lora_rank * 32 * 2 * 128 * 2 / 1024;
     let estimated_memory_mb = model_memory + lora_memory + 500;
 
     AutoTuneConfig {
